New Palette Fix
For some reason the new palette stuff that were drafted never got added for unknown reasons, and I just noticed this today while playing RDV lol. Verified that I wasn't insane and checked the releases and the repo and yep, they didn't get ever updated to be the newer ones.